General Information Half Point Complete
medical users can request to add research to a general info tab that all but admin can see, and admin must approve/reject request and it will or will not show up depending on approved or not. Also, it organizes so the ~most recent~ shows up first, and doctors put down the hospital they do research with cause it could be different than where they work, I don't know how that works I'm not a doctor.